fix(core): cleanup `rxResource` abort listener
This commit updates the `rxResource` functionality. Currently, we're attaching an `abort` event
listener to the signal (when provided) and never removing it. To ensure proper resource management,
we should remove event listeners when they are no longer needed, allowing objects to be garbage
collected. In Firefox, saving a heap snapshot and analyzing it with `fxsnapshot` reveals that querying
`AbortSignal` prints a "CaptureMap" with a list of "lambdas", indicating that the signal is not
garbage collected due to the event listener's lambda function.