feat(platform-browser): add Hammer provider functions
Prior to this commit, there were no functions that allowed providing HammerJS in a "new" way through environment providers; it required using `importProvidersFrom(HammerModule)` when bootstrapping the application. In this commit, we add two new functions that return environment providers and can be called directly in the providers when bootstrapping the application.